Move splashPage into Pages directory
